+++ tests/test_plot/test_misc.py
@@ -749,6 +749,18 @@ def check_makefile(ctx, mkfile, prj, dbg, txt):
     assert board_file in deps
     assert yaml in deps
     logging.debug('- Target `Board View Test` OK')
+    # pcb_stats
+    if context.ki10():
+        deps = targets['pcb_stats'].split(' ')
+        assert len(deps) == 1, deps
+        fname = ctx.get_out_path(prj+'-statistics.txt')
+        assert fname in deps
+        deps = targets[targets['pcb_stats']].split(' ')
+        assert len(deps) == 2
+        assert board_file in deps
+        assert yaml in deps
+        ctx.expect_out_file(os.path.basename(fname))
+        logging.debug('- Target `pcb_stats` OK')
     # pdf_sch_int
     deps = targets['pdf_sch_int'].split(' ')
     assert len(deps) == 1, deps
@@ -809,7 +821,7 @@ def check_makefile(ctx, mkfile, prj, dbg, txt):
     assert len(deps) == 1, deps
     assert ctx.get_out_path(prj+'-archive.zip') in deps
     deps = targets[targets['archive']].split(' ')
-    assert len(deps) == 18, deps
+    assert len(deps) == 19, deps
     # - position
     assert ctx.get_out_path(os.path.join(POS_DIR, prj+'-top_pos.csv')) in deps
     assert ctx.get_out_path(os.path.join(POS_DIR, prj+'-bottom_pos.csv')) in deps
@@ -842,9 +854,9 @@ def check_makefile(ctx, mkfile, prj, dbg, txt):
 @pytest.mark.kicad2step
 def test_makefile_1(test_dir):
     prj = 'test_v5'
-    ctx = context.TestContext(test_dir, prj, 'makefile_1')
+    ctx = context.TestContext(test_dir, prj, 'makefile_1_ki10' if context.ki10() else 'makefile_1')
     mkfile = ctx.get_out_path('Makefile')
-    ctx.run(extra=['-s', 'all', 'archive'])
+    ctx.run(extra=['-s', 'all', 'archive', 'pcb_stats'])
     ctx.run(extra=['-m', mkfile])
     check_makefile(ctx, mkfile, prj, '-v', r"^\t\$\(KIBOT_CMD\) -s (.*) -i$")
     ctx.clean_up()
@@ -854,9 +866,9 @@ def test_makefile_1(test_dir):
 @pytest.mark.kicad2step
 def test_makefile_2(test_dir):
     prj = 'test_v5'
-    ctx = context.TestContext(test_dir, prj, 'makefile_1')
+    ctx = context.TestContext(test_dir, prj, 'makefile_1_ki10' if context.ki10() else 'makefile_1')
     mkfile = ctx.get_out_path('Makefile')
-    ctx.run(extra=['-s', 'all', 'archive'])
+    ctx.run(extra=['-s', 'all', 'archive', 'pcb_stats'])
     ctx.run(extra=['-m', mkfile], no_verbose=True)
     check_makefile(ctx, mkfile, prj, '', r"^\t@\$\(KIBOT_CMD\) -s (.*) -i 2>> \$\(LOGFILE\)$")
     ctx.clean_up()