+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - *SCH Print:
   - The list of targets now includes all schematic pages
   - Output when the `output` option was empty
+- PCB Stats:
+  - Missing targets
 - Report:
   - Now all SVG schematic pages are included, not just the first